multi vendor conversion

+++ backend/customer/helpers/mikrotik.py
@@ -18,21 +18,30 @@ class Mikrotik:
     """
 
     @staticmethod
-    def get_user_by_username(username):
+    def get_user_by_username(username, organization=None):
         """
         Fetch a PPP secret by username.
         :param username: str
         :return: tuple(success: bool, data_or_error: dict/str)
         """
+        if (
+            not organization
+            or not organization.router_ip
+            or not organization.router_username
+            or not organization.router_password
+        ):
+            print("Router configuration is not added yet")
+            return False, f"Router configuration is not added yet"
+
         if not username:
             return False, "Username is required"
 
         try:
-            query_url = f"{MIKROTIK_URL}/rest/ppp/secret/print"
+            query_url = f"{organization.router_ip}/rest/ppp/secret/print"
             response = requests.post(
                 query_url,
                 json={".query": [f"name={username}"]},
-                auth=(MIKROTIK_USER, MIKROTIK_PASS),
+                auth=(organization.router_username, organization.router_password),
                 verify=False,
                 # verify=MIKROTIK_VERIFY_SSL,
                 # timeout=REQUEST_TIMEOUT,
@@ -55,16 +64,24 @@ class Mikrotik:
             return False, f"Unexpected error: {str(e)}"
 
     @staticmethod
-    def get_users_from_server():
+    def get_users_from_server(organization=None):
         """
         Fetch all PPP users (secrets).
         :return: tuple(success: bool, users_list_or_error: list/str)
         """
+        if (
+            not organization
+            or not organization.router_ip
+            or not organization.router_username
+            or not organization.router_password
+        ):
+            print("Router configuration is not added yet")
+            return False, f"Router configuration is not added yet"
         try:
-            url = f"{MIKROTIK_URL}/rest/ppp/secret"
+            url = f"{organization.router_ip}/rest/ppp/secret"
             response = requests.get(
                 url,
-                auth=(MIKROTIK_USER, MIKROTIK_PASS),
+                auth=(organization.router_username, organization.router_password),
                 # verify=MIKROTIK_VERIFY_SSL,
                 # timeout=REQUEST_TIMEOUT,
                 verify=False,
@@ -83,25 +100,32 @@ class Mikrotik:
             return False, f"Unexpected error: {str(e)}"
 
     @staticmethod
-    def get_user_sessions():
+    def get_user_sessions(organization):
         """
         Get all active PPP sessions.
         :return: tuple(success: bool, sessions_list_or_error: list/str)
         """
+        if (
+            not organization.router_ip
+            or not organization.router_username
+            or not organization.router_password
+        ):
+            print("Router configuration is not added yet")
+            return False, []
         try:
-            url = f"{MIKROTIK_URL}/rest/ppp/active"
+            url = f"{organization.router_ip}/rest/ppp/active"
             response = requests.get(
                 url,
-                auth=(MIKROTIK_USER, MIKROTIK_PASS),
+                auth=(organization.router_username, organization.router_password),
                 # verify=MIKROTIK_VERIFY_SSL,
                 # timeout=REQUEST_TIMEOUT,
                 verify=False,
             )
-
+            # print("RRRRR: ", response)
             if response.status_code == 200:
                 return True, response.json()
-            else:
-                return False, f"HTTP {response.status_code}: {response.text}"
+
+            return False, []
 
         except requests.exceptions.Timeout:
             return False, "Request timed out"
@@ -111,17 +135,24 @@ class Mikrotik:
             return False, f"Unexpected error: {str(e)}"
 
     @staticmethod
-    def delete_user_session(session_id):
+    def delete_user_session(session_id, organization=None):
         """
         Terminate an active PPP session by ID.
         :param session_id: str (.id from session)
         :return: bool
         """
+        if (
+            not organization.router_ip
+            or not organization.router_username
+            or not organization.router_password
+        ):
+            print("Router configuration is not added yet")
+            return False
         try:
-            url = f"{MIKROTIK_URL}/rest/ppp/active/{session_id}"
+            url = f"{organization.router_ip}/rest/ppp/active/{session_id}"
             response = requests.delete(
                 url,
-                auth=(MIKROTIK_USER, MIKROTIK_PASS),
+                auth=(organization.router_username, organization.router_password),
                 # verify=MIKROTIK_VERIFY_SSL,
                 # timeout=REQUEST_TIMEOUT,
                 verify=False,
@@ -140,7 +171,7 @@ class Mikrotik:
             return False
 
     @staticmethod
-    def toggle_ppp_user(username, disable=True):
+    def toggle_ppp_user(username, disable=True, organization=None, session_id=None):
         """
         Enable or disable a PPP user.
         If disabling, also terminate active session.
@@ -151,9 +182,17 @@ class Mikrotik:
         if not username:
             return False, "Username is required"
 
+        if (
+            not organization.router_ip
+            or not organization.router_username
+            or not organization.router_password
+        ):
+            print("Router configuration is not added yet")
+            return False, f"Router configuration is not added yet"
+
         try:
             # Step 1: Get user
-            success, user = Mikrotik.get_user_by_username(username)
+            success, user = Mikrotik.get_user_by_username(username, organization)
             if not success:
                 return False, f"User not found: {user}"
 
@@ -161,11 +200,11 @@ class Mikrotik:
             disabled_str = "true" if disable else "false"
 
             # Step 2: Update disabled status
-            patch_url = f"{MIKROTIK_URL}/rest/ppp/secret/{secret_id}"
+            patch_url = f"{organization.router_ip}/rest/ppp/secret/{secret_id}"
             response = requests.patch(
                 patch_url,
                 json={"disabled": disabled_str},
-                auth=(MIKROTIK_USER, MIKROTIK_PASS),
+                auth=(organization.router_username, organization.router_password),
                 # verify=MIKROTIK_VERIFY_SSL,
                 # timeout=REQUEST_TIMEOUT,
                 verify=False,
@@ -179,8 +218,15 @@ class Mikrotik:
                 return False, f"Update failed: {error_msg}"
 
             # Step 3: If disabling, terminate active session
-            if disable:
-                success, sessions = Mikrotik.get_user_sessions()
+            if disable and session_id:
+                if Mikrotik.delete_user_session(session_id):
+                    print(f"[Mikrotik] Terminated session for '{username}'")
+                else:
+                    print(
+                        f"[Mikrotik] Warning: Could not terminate session for '{username}'"
+                    )
+            elif disable:
+                success, sessions = Mikrotik.get_user_sessions(organization)
                 if success:
                     for session in sessions:
                         if session.get("name") == username:
@@ -191,7 +237,7 @@ class Mikrotik:
                                 print(
                                     f"[Mikrotik] Warning: Could not terminate session for '{username}'"
                                 )
-                    print("sessions not available for the user")
+                    # print("sessions not available for the user")
                 else:
                     print(
                         f"[Mikrotik] Warning: Could not fetch active sessions: {sessions}"
@@ -208,7 +254,16 @@ class Mikrotik:
             return False, f"Unexpected error: {str(e)}"
 
     @staticmethod
-    def create_ppp_user(user):
+    def create_ppp_user(user, organization=None):
+        if (
+            not organization
+            or not organization.router_ip
+            or not organization.router_username
+            or not organization.router_password
+        ):
+            print("Router configuration is not added yet")
+            return False, f"Router configuration is not added yet"
+
         username = user.get("username", "")
         password = user.get("password", "")
         if not username or not password:
@@ -218,30 +273,36 @@ class Mikrotik:
             "name": username,
             "password": password,
             "service": user.get("service", "pppoe"),  # Default to pppoe
-            "profile": user.get("profile", "5Mbps"),  # Default profile
+            "profile": user.get("profile", "default"),  # Default profile
             "disabled": "false",
-            "comment": user.get("comment", "Test Comment"),
+            "comment": user.get("comment", "Created via API"),
         }
-
+        # print("Payload to create PPP user: ", payload)
         try:
-            url = f"{MIKROTIK_URL}/rest/ppp/secret"
+            url = f"{organization.router_ip}/rest/ppp/secret"
             response = requests.put(  # MikroTik uses PUT to CREATE
                 url,
                 json=payload,
-                auth=(MIKROTIK_USER, MIKROTIK_PASS),
+                auth=(organization.router_username, organization.router_password),
                 verify=False,
                 # verify=MIKROTIK_VERIFY_SSL,
                 # timeout=REQUEST_TIMEOUT,
             )
-            print("RRRRR: ", response)
+            # print("RRRRR: ", response)
             if response.status_code == 201:
                 return True, "PPP user created successfully"
             else:
                 try:
-                    error_msg = response.json().get("message", response.text)
+                    # print("Response JSON: ", response.json())
+                    response = response.json()
+                    return (
+                        False,
+                        f"{response.get('detail', 'Failed to create user')}",
+                    )
+
                 except:
-                    error_msg = response.text
-                return False, f"Create failed [{response.status_code}]: {error_msg}"
+                    # print("Response Text: ", response.text)
+                    return False, f"Create failed: {response.text}"
 
         except requests.exceptions.Timeout:
             return False, "Request timed out"
@@ -251,30 +312,37 @@ class Mikrotik:
             return False, f"Unexpected error: {str(e)}"
 
     @staticmethod
-    def delete_ppp_user(username):
+    def delete_ppp_user(username, organization=None):
         """
         Delete a PPP user (ppp secret) by username.
         :param username: str
         :return: tuple(success: bool, message: str)
         """
-        success, user = Mikrotik.get_user_by_username(username)
+        if (
+            not organization
+            or not organization.router_ip
+            or not organization.router_username
+            or not organization.router_password
+        ):
+            print("Router configuration is not added yet")
+            return False, f"Router configuration is not added yet"
+
+        success, user = Mikrotik.get_user_by_username(username, organization)
         if not success:
             return False, f"User not found: {user}"
 
         try:
-            url = f"{MIKROTIK_URL}/rest/ppp/secret/{user['.id']}"
+            url = f"{organization.router_ip}/rest/ppp/secret/{user['.id']}"
             response = requests.delete(
                 url,
-                auth=(MIKROTIK_USER, MIKROTIK_PASS),
+                auth=(organization.router_username, organization.router_password),
                 verify=False,
                 # verify=MIKROTIK_VERIFY_SSL,
                 # timeout=REQUEST_TIMEOUT,
             )
-
-            if response.status_code == 200:
-                return True, "User deleted successfully"
-            else:
-                return False, f"Delete failed: {response.text}"
+            if response.status_code != 204:
+                return False, f"Failed to delete user: {response.text}"
+            return True, "User deleted successfully"
 
         except requests.exceptions.Timeout:
             return False, "Request timed out"
